requirements.txt
# FastAPI Backend Requirements
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn==21.2.0

# Database
sqlalchemy[asyncio]==2.0.43
asyncpg==0.30.0
alembic==1.16.5

# Data processing
pandas==2.3.2
numpy==2.3.3

# AI Integration
openai==1.107.1

# Web Framework
starlette==0.47.3
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic[email]==2.11.7

# HTTP Client
httpx==0.28.1

# Security
itsdangerous==2.2.0

# Development
requests==2.32.3